requirements1.txt
# Core
streamlit
numpy
pandas
matplotlib
seaborn
pillow
requests
textblob

# Machine Learning
scikit-learn
xgboost
tensorflow==2.13.1
keras==2.13.1

# Agent-Based Modeling
mesa==2.1.5

